% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALARM-package.R
\docType{package}
\name{ALARM-package}
\alias{ALARM}
\alias{ALARM-package}
\title{ALARM: Asian Lung Cancer Absolute Risk Models}
\description{
This package provides a user-friendly API for making predictions using the Asian Lung Cancer Absolute Risk Models (ALARM) for lung cancer morality as published in Warkentin et al. (2022). Read more about these models at https://doi.org/10.1093/jnci/djac176.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattwarkentin/ALARM}
  \item Report bugs at \url{https://github.com/mattwarkentin/ALARM/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{matthew.warkentin@ucalgary.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID})

}
\keyword{internal}
