% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{alarm}
\alias{alarm}
\title{ALARM Project Color Palettes}
\format{
A list of character vectors of hexcodes for palettes.
}
\usage{
alarm
}
\description{
ALARM Project Color Palettes
}
\examples{
alarm$alarm

}
\keyword{datasets}
