% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{alarm_ppt_create}
\alias{alarm_ppt_create}
\title{Create a PowerPoint with ALARM Project theming}
\usage{
alarm_ppt_create(template = "light")
}
\arguments{
\item{template}{character string, either \code{'light'} or \code{'dark'}}
}
\value{
an object of class \code{rpptx}
}
\description{
Create a PowerPoint with ALARM Project theming
}
\examples{
file_ppt <- tempfile(fileext = '.pptx')
alarm_ppt_create() |>
  print(target = file_ppt)
}
