% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{alarm_ppt_template}
\alias{alarm_ppt_template}
\title{Get ALARM PowerPoint templates}
\usage{
alarm_ppt_template(template = "light")
}
\arguments{
\item{template}{character string, either \code{'light'} or \code{'dark'}}
}
\value{
path to a \code{'.potx'} file
}
\description{
Get ALARM PowerPoint templates
}
\examples{
alarm_ppt_template('light')
}
