% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{gg_annotate_alarm}
\alias{gg_annotate_alarm}
\title{Add ALARM Project logo to a ggplot2 object}
\usage{
gg_annotate_alarm(gg)
}
\arguments{
\item{gg}{a ggplot2 object}
}
\value{
a ggplot2 object
}
\description{
Add ALARM Project logo to a ggplot2 object
}
\examples{
p <- ggplot2::mpg |>
  ggplot2::ggplot() +
  ggplot2::geom_point(ggplot2::aes(displ, hwy, color = class)) +
  scale_color_alarm() +
  theme_alarm_dark()

p |> gg_annotate_alarm()
}
