% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_alarm}
\alias{theme_alarm}
\alias{theme_alarm_dark}
\alias{theme_alarm_light}
\title{ALARM Project themes for \link{ggplot2}}
\usage{
theme_alarm_dark(...)

theme_alarm_light(...)
}
\arguments{
\item{...}{arguments to pass to \link[ggplot2:theme]{ggplot2::theme}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
ALARM Project themes for \link{ggplot2}
}
\examples{
ggplot2::mpg |>
  ggplot2::ggplot() +
   ggplot2::geom_point(ggplot2::aes(displ, hwy, color = class)) +
   scale_color_alarm() +
   theme_alarm_light()

ggplot2::mpg |>
  ggplot2::ggplot() +
   ggplot2::geom_point(ggplot2::aes(displ, hwy, color = class)) +
   ggplot2::facet_wrap(~manufacturer) +
   scale_color_alarm() +
   theme_alarm_light()

ggplot2::mpg |>
  ggplot2::ggplot() +
   ggplot2::geom_point(ggplot2::aes(displ, hwy, color = class)) +
   ggplot2::facet_wrap(~manufacturer) +
   scale_color_alarm() +
   theme_alarm_dark()

}
