% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{gg_annotate_alarm}
\alias{gg_annotate_alarm}
\title{Add ALARM Project logo to a ggplot2 object}
\usage{
gg_annotate_alarm(
  gg,
  logo_left = 0.9,
  logo_bottom = 0.01,
  logo_right = 0.99,
  logo_top = 0.1,
  band = TRUE
)
}
\arguments{
\item{gg}{a ggplot2 object}

\item{logo_left}{relative position for logo from the left}

\item{logo_bottom}{relative position for logo from the bottom}

\item{logo_right}{relative position for logo from the right}

\item{logo_top}{relative position for logo from the top}

\item{band}{logical, default \code{TRUE}. Should the logo be placed on a blue band?}
}
\value{
a ggplot2 object
}
\description{
Add ALARM Project logo to a ggplot2 object
}
\examples{
p <- ggplot2::mpg |>
  ggplot2::ggplot() +
  ggplot2::geom_point(ggplot2::aes(displ, hwy, color = class)) +
  scale_color_alarm() +
  theme_alarm_dark()

p |> gg_annotate_alarm()
}
