% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_data}
\alias{validate_data}
\title{Validate Data for ALARM}
\usage{
validate_data(data)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} with the following numeric columns
(in any order):
\itemize{
\item \code{age} - Age (years)
\item \code{sex} - Sex (0 = Man, 1 = Woman)
\item \code{fhx_cancer} - Family history of any cancer? (0 = No, 1 = Yes)
\item \code{phx_cancer} - Personal history of cancer? (0 = No, 1 = Yes)
\item \code{fev1fvc} - FEV1 / FVC (between 0 and 1)
\item \code{phx_lungdx} - Personal history of COPD? (0 = No, 1 = Yes)
\item \code{hhinc} - Household income (see \code{Details})
\item \code{bmi} - Body Mass Index (kg/m^2)
\item \code{smk_ever} - Ever smoker? (0 = No/Never, 1 = Yes/Ever)
\item \code{smk_former} - Former smoker? (0 = No, 1 = Yes)
\item \code{smk_duration} - Smoking duration (years)
\item \code{smk_cigpday} - Smoking intensity (no. cigarettes per day)
\item \code{smk_years_stop} - Years since quitting (years)
}}
}
\value{
Logical. \code{TRUE} if the data set is valid, \code{FALSE} otherwise.
}
\description{
\code{validate_data} is used to ensure all of the predictors are in the expected
format before making predictions. This function is called internally by
\code{predict.ALARM()}.
}
\details{
Household income has 6 levels. The levels are defined in yuan, and
the approximate USD equivalents are present in parentheses. Household
income levels: \verb{0 = < 2500} (< 384), \code{1 = 2500-4999} (384-767),
\code{2 = 5000-9999} (767-1537), \code{3 = 10000-19999} (1537-3075),
\code{4 = 20000-34999} (3075-5381), and \verb{5 = >= 35000} (>= 5381).
}
